+++ backend/api/validation/user.go
@@ -1,4 +1,4 @@
-package validator
+package validation
 
 import (
 	"incomster/backend/api/oas"
